// projects/vio-dashboard/src/lib/models/dashboard.models.ts
export interface ChartSegment {
  label: string;
  value: number;
  color: string;
}